Ask method implemented
If a boolean value is asked, then a boolean value is returned depending on if the character has that characteristic. If a non-boolean value is asked, and the character has that characteristic, then the name of that characteristic is returned. If a non-boolean value is asked, and the character does NOT have that characteristic, then the phrase "not " is prepended to the name of that characteristic before returning that value.