Ask_vars method implemented
Returns a list of characters that matches the given characteristic. The value for the characteristic can either be boolean, or non-boolean. If non-boolean, then special care must be taken to remove a prepended "not " if it exists. Furthermore, if a "not " is used, then the check must be inverted. Instead of finding every character that matches the characteristic, it finds every character that does NOT match the characteristic.